Modified the OSMStateManager into a builder pattern so that it can only be instanced once and is accessed everywhere
